Update code and utilities
